client: split client library by API (& legacy)
Split the client library by the API and separate from legacy API code
and core client code.
