+++ client/client.go
@@ -25,15 +25,12 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"os"
 	"strings"
-	"time"
 
 	"github.com/ciao-project/ciao/ciao-controller/api"
 	"github.com/ciao-project/ciao/ciao-controller/types"
-	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/pkg/errors"
 )
 
@@ -225,8 +222,6 @@ func (client *Client) sendHTTPRequest(method string, url string, values []queryV
 }
 
 func (client *Client) unmarshalHTTPResponse(resp *http.Response, v interface{}) error {
-	defer resp.Body.Close()
-
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		return errors.Wrap(err, "Could not read HTTP response body")
@@ -358,958 +353,4 @@ func (client *Client) postResource(url string, content string, request interface
 	}
 
 	return nil
-}
-
-// ListEvents retrieves the events for either all or the desired tenant
-func (client *Client) ListEvents(tenantID string) (types.CiaoEvents, error) {
-	var events types.CiaoEvents
-	var url string
-
-	if tenantID == "" {
-		url = client.buildComputeURL("events")
-	} else {
-		url = client.buildComputeURL("%s/events", tenantID)
-	}
-
-	err := client.getResource(url, "", nil, &events)
-
-	return events, err
-}
-
-// DeleteEvents deletes all events from
-func (client *Client) DeleteEvents() error {
-	url := client.buildComputeURL("events")
-
-	return client.deleteResource(url, "")
-}
-
-func (client *Client) getCiaoExternalIPsResource() (string, string, error) {
-	url, err := client.getCiaoResource("external-ips", api.ExternalIPsV1)
-	return url, api.ExternalIPsV1, err
-}
-
-// TBD: in an ideal world, we'd modify the GET to take a query.
-func (client *Client) getExternalIPRef(address string) (string, error) {
-	var IPs []types.MappedIP
-
-	url, ver, err := client.getCiaoExternalIPsResource()
-	if err != nil {
-		return "", err
-	}
-
-	err = client.getResource(url, ver, nil, &IPs)
-	if err != nil {
-		return "", err
-	}
-
-	for _, IP := range IPs {
-		if IP.ExternalIP == address {
-			url := client.getRef("self", IP.Links)
-			if url != "" {
-				return url, nil
-			}
-		}
-	}
-
-	return "", types.ErrAddressNotFound
-}
-
-// MapExternalIP maps an IP from the pool to the given instance
-func (client *Client) MapExternalIP(pool string, instanceID string) error {
-	req := types.MapIPRequest{
-		InstanceID: instanceID,
-	}
-
-	if pool != "" {
-		req.PoolName = &pool
-	}
-
-	url, ver, err := client.getCiaoExternalIPsResource()
-	if err != nil {
-		return errors.Wrap(err, "Error getting external IP resource")
-	}
-
-	return client.postResource(url, ver, &req, nil)
-}
-
-// ListExternalIPs returns the mapped IPs
-func (client *Client) ListExternalIPs() ([]types.MappedIP, error) {
-	var IPs []types.MappedIP
-
-	url, ver, err := client.getCiaoExternalIPsResource()
-	if err != nil {
-		return IPs, errors.Wrap(err, "Error getting external IP resource")
-	}
-
-	err = client.getResource(url, ver, nil, &IPs)
-
-	return IPs, err
-}
-
-// UnmapExternalIP unmaps the given address from the instance
-func (client *Client) UnmapExternalIP(address string) error {
-	url, err := client.getExternalIPRef(address)
-	if err != nil {
-		return errors.Wrap(err, "Error getting external IP reference")
-	}
-
-	return client.deleteResource(url, api.ExternalIPsV1)
-}
-
-func (client *Client) getCiaoPoolsResource() (string, error) {
-	return client.getCiaoResource("pools", api.PoolsV1)
-}
-
-func (client *Client) getCiaoPoolRef(name string) (string, error) {
-	var pools types.ListPoolsResponse
-
-	query := queryValue{
-		name:  "name",
-		value: name,
-	}
-
-	url, err := client.getCiaoPoolsResource()
-	if err != nil {
-		return "", err
-	}
-
-	err = client.getResource(url, api.PoolsV1, []queryValue{query}, &pools)
-	if err != nil {
-		return "", err
-	}
-
-	// we have now the pool ID
-	if len(pools.Pools) != 1 {
-		return "", errors.New("No pool by that name found")
-	}
-
-	links := pools.Pools[0].Links
-	url = client.getRef("self", links)
-	if url == "" {
-		return url, errors.New("Invalid Link returned from controller")
-	}
-
-	return url, nil
-}
-
-// GetExternalIPPool gets the details of a single external IP pool
-func (client *Client) GetExternalIPPool(name string) (types.Pool, error) {
-	var pool types.Pool
-
-	if !client.IsPrivileged() {
-		return pool, errors.New("This command is only available to admins")
-	}
-
-	url, err := client.getCiaoPoolRef(name)
-	if err != nil {
-		return pool, err
-	}
-
-	err = client.getResource(url, api.PoolsV1, nil, &pool)
-	return pool, err
-}
-
-// CreateExternalIPPool creates a pool of IPs
-func (client *Client) CreateExternalIPPool(name string) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	req := types.NewPoolRequest{
-		Name: name,
-	}
-
-	url, err := client.getCiaoPoolsResource()
-	if err != nil {
-		return errors.Wrap(err, "Error getting pool resource")
-	}
-
-	return client.postResource(url, api.PoolsV1, &req, nil)
-}
-
-// ListExternalIPPools lists the pools in which IPs are available
-func (client *Client) ListExternalIPPools() (types.ListPoolsResponse, error) {
-	var pools types.ListPoolsResponse
-
-	url, err := client.getCiaoPoolsResource()
-	if err != nil {
-		return pools, errors.Wrap(err, "Error getting pool resource")
-	}
-
-	err = client.getResource(url, api.PoolsV1, nil, &pools)
-
-	return pools, err
-}
-
-// DeleteExternalIPPool deletes the pool of the given name
-func (client *Client) DeleteExternalIPPool(pool string) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	url, err := client.getCiaoPoolRef(pool)
-	if err != nil {
-		return errors.Wrap(err, "Error getting pool reference")
-	}
-
-	return client.deleteResource(url, api.PoolsV1)
-
-}
-
-// AddExternalIPSubnet adds a subnet to the external IP pool
-func (client *Client) AddExternalIPSubnet(pool string, subnet *net.IPNet) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	var req types.NewAddressRequest
-
-	url, err := client.getCiaoPoolRef(pool)
-	if err != nil {
-		return errors.Wrap(err, "Error getting pool reference")
-	}
-
-	s := subnet.String()
-	req.Subnet = &s
-
-	return client.postResource(url, api.PoolsV1, &req, nil)
-}
-
-// AddExternalIPAddresses adds a set of IP addresses to the external IP pool
-func (client *Client) AddExternalIPAddresses(pool string, IPs []string) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	var req types.NewAddressRequest
-
-	url, err := client.getCiaoPoolRef(pool)
-	if err != nil {
-		return errors.Wrap(err, "Error getting pool reference")
-	}
-
-	for _, IP := range IPs {
-		addr := types.NewIPAddressRequest{
-			IP: IP,
-		}
-
-		req.IPs = append(req.IPs, addr)
-	}
-
-	return client.postResource(url, api.PoolsV1, &req, nil)
-}
-
-func (client *Client) getSubnetRef(pool types.Pool, cidr string) string {
-	for _, sub := range pool.Subnets {
-		if sub.CIDR == cidr {
-			return client.getRef("self", sub.Links)
-		}
-	}
-
-	return ""
-}
-
-func (client *Client) getIPRef(pool types.Pool, address string) string {
-	for _, ip := range pool.IPs {
-		if ip.Address == address {
-			return client.getRef("self", ip.Links)
-		}
-	}
-
-	return ""
-}
-
-// RemoveExternalIPSubnet removes a subnet from the pool
-func (client *Client) RemoveExternalIPSubnet(pool string, subnet *net.IPNet) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	p, err := client.GetExternalIPPool(pool)
-	if err != nil {
-		return errors.Wrap(err, "Error getting external IP pool")
-	}
-
-	url := client.getSubnetRef(p, subnet.String())
-	if url == "" {
-		return fmt.Errorf("Subnet not present in pool")
-	}
-
-	return client.deleteResource(url, api.PoolsV1)
-}
-
-// RemoveExternalIPAddress removes a single IP address from the pool
-func (client *Client) RemoveExternalIPAddress(pool string, IP string) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	p, err := client.GetExternalIPPool(pool)
-	if err != nil {
-		return errors.Wrap(err, "Error getting external IP pool")
-	}
-
-	url := client.getIPRef(p, IP)
-	if url == "" {
-		return fmt.Errorf("IP not present in pool")
-	}
-
-	return client.deleteResource(url, api.PoolsV1)
-}
-
-// GetImage retrieves the details for an image
-func (client *Client) GetImage(imageID string) (types.Image, error) {
-	var i types.Image
-
-	var url string
-	if client.IsPrivileged() && client.TenantID == "admin" {
-		url = client.buildCiaoURL("images/%s", imageID)
-	} else {
-		url = client.buildCiaoURL("%s/images/%s", client.TenantID, imageID)
-	}
-
-	err := client.getResource(url, api.ImagesV1, nil, &i)
-
-	return i, err
-}
-
-func (client *Client) uploadTenantImage(tenant, image string, data io.Reader) error {
-	var url string
-	if client.IsPrivileged() && client.TenantID == "admin" {
-		url = client.buildCiaoURL("images/%s/file", image)
-	} else {
-		url = client.buildCiaoURL("%s/images/%s/file", client.TenantID, image)
-	}
-
-	resp, err := client.sendHTTPRequest("PUT", url, nil, data, fmt.Sprintf("%s/octet-stream", api.ImagesV1))
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusNoContent {
-		return fmt.Errorf("Unexpected HTTP response code (%d): %s", resp.StatusCode, resp.Status)
-	}
-
-	return err
-}
-
-// CreateImage creates and uploads a new image
-func (client *Client) CreateImage(name string, visibility types.Visibility, ID string, data io.Reader) (string, error) {
-	opts := api.CreateImageRequest{
-		Name:       name,
-		ID:         ID,
-		Visibility: visibility,
-	}
-
-	var url string
-	if client.IsPrivileged() && client.TenantID == "admin" {
-		url = client.buildCiaoURL("images")
-	} else {
-		url = client.buildCiaoURL("%s/images", client.TenantID)
-	}
-
-	var image types.Image
-	err := client.postResource(url, api.ImagesV1, &opts, &image)
-	if err != nil {
-		return "", errors.Wrap(err, "Error creating image resource")
-	}
-
-	err = client.uploadTenantImage(client.TenantID, image.ID, data)
-	if err != nil {
-		return "", errors.Wrap(err, "Error uploading image data")
-	}
-
-	return image.ID, nil
-}
-
-// ListImages retrieves the set of available images
-func (client *Client) ListImages() ([]types.Image, error) {
-	var images []types.Image
-
-	var url string
-	if client.IsPrivileged() && client.TenantID == "admin" {
-		url = client.buildCiaoURL("images")
-	} else {
-		url = client.buildCiaoURL("%s/images", client.TenantID)
-	}
-
-	err := client.getResource(url, api.ImagesV1, nil, &images)
-
-	return images, err
-}
-
-// DeleteImage deletes the given image
-func (client *Client) DeleteImage(imageID string) error {
-	var url string
-	if client.IsPrivileged() && client.TenantID == "admin" {
-		url = client.buildCiaoURL("images/%s", imageID)
-	} else {
-		url = client.buildCiaoURL("%s/images/%s", client.TenantID, imageID)
-	}
-
-	return client.deleteResource(url, api.ImagesV1)
-}
-
-// CreateInstances creates instances by the given request
-func (client *Client) CreateInstances(request api.CreateServerRequest) (api.Servers, error) {
-	var servers api.Servers
-
-	url := client.buildCiaoURL("%s/instances", client.TenantID)
-	err := client.postResource(url, api.InstancesV1, &request, &servers)
-
-	return servers, err
-}
-
-// DeleteInstance deletes the given instance
-func (client *Client) DeleteInstance(instanceID string) error {
-	url := client.buildCiaoURL("%s/instances/%s", client.TenantID, instanceID)
-	return client.deleteResource(url, api.InstancesV1)
-}
-
-func (client *Client) instanceAction(instanceID string, action string) error {
-	actionBytes := []byte(action)
-
-	url := client.buildCiaoURL("%s/instances/%s/action", client.TenantID, instanceID)
-
-	resp, err := client.sendHTTPRequest("POST", url, nil, bytes.NewReader(actionBytes), api.InstancesV1)
-	if err != nil {
-		return errors.Wrap(err, "Error making HTTP request")
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusAccepted {
-		return fmt.Errorf("HTTP response code from %s not as expected: %d", url, resp.StatusCode)
-	}
-	return nil
-}
-
-// StopInstance stops the given instance
-func (client *Client) StopInstance(instanceID string) error {
-	return client.instanceAction(instanceID, "os-stop")
-}
-
-// StartInstance stops the given instance
-func (client *Client) StartInstance(instanceID string) error {
-	return client.instanceAction(instanceID, "os-start")
-}
-
-// ListInstancesByWorkload provides the list of instances for a given tenant and workloadID.
-func (client *Client) ListInstancesByWorkload(tenantID string, workloadID string) (api.Servers, error) {
-	var servers api.Servers
-
-	url := client.buildCiaoURL("%s/instances/detail", tenantID)
-
-	values := []queryValue{}
-	if workloadID != "" {
-		values = append(values, queryValue{
-			name:  "workload",
-			value: workloadID,
-		})
-	}
-
-	err := client.getResource(url, api.InstancesV1, values, &servers)
-
-	return servers, err
-
-}
-
-// ListInstances gets the set of instances
-func (client *Client) ListInstances() (api.Servers, error) {
-	return client.ListInstancesByWorkload(client.TenantID, "")
-}
-
-// GetInstance gets the details of a single instances
-func (client *Client) GetInstance(instanceID string) (api.Server, error) {
-	var server api.Server
-
-	url := client.buildCiaoURL("%s/instances/%s", client.TenantID, instanceID)
-	err := client.getResource(url, api.InstancesV1, nil, &server)
-
-	return server, err
-}
-
-// ListInstancesByNode gets the instances on a given node
-func (client *Client) ListInstancesByNode(nodeID string) (types.CiaoServersStats, error) {
-	var servers types.CiaoServersStats
-
-	url := client.buildComputeURL("nodes/%s/servers/detail", nodeID)
-	err := client.getResource(url, "", nil, &servers)
-
-	return servers, err
-}
-
-// DeleteAllInstances deletes all the instances
-func (client *Client) DeleteAllInstances() error {
-	var action types.CiaoServersAction
-
-	url := client.buildComputeURL("%s/servers/action", client.TenantID)
-	action.Action = "os-delete"
-
-	return client.postResource(url, "", &action, nil)
-}
-
-func (client *Client) getCiaoWorkloadsResource() (string, error) {
-	return client.getCiaoResource("workloads", api.WorkloadsV1)
-}
-
-// ListWorkloads gets the workloads available
-func (client *Client) ListWorkloads() ([]types.Workload, error) {
-	var wls []types.Workload
-
-	var url string
-	if client.IsPrivileged() {
-		url = client.buildCiaoURL("workloads")
-	} else {
-		url = client.buildCiaoURL("%s/workloads", client.TenantID)
-	}
-
-	err := client.getResource(url, api.WorkloadsV1, nil, &wls)
-	return wls, err
-}
-
-// CreateWorkload creates a worklaod
-func (client *Client) CreateWorkload(request types.Workload) (string, error) {
-	url, err := client.getCiaoWorkloadsResource()
-	if err != nil {
-		return "", errors.Wrap(err, "Error getting workloads resource")
-	}
-
-	var response types.WorkloadResponse
-
-	err = client.postResource(url, api.WorkloadsV1, &request, &response)
-
-	return response.Workload.ID, err
-}
-
-// DeleteWorkload deletes the given workload
-func (client *Client) DeleteWorkload(workloadID string) error {
-	url, err := client.getCiaoWorkloadsResource()
-	if err != nil {
-		return errors.Wrap(err, "Error getting workloads resource")
-	}
-
-	url = fmt.Sprintf("%s/%s", url, workloadID)
-
-	return client.deleteResource(url, api.WorkloadsV1)
-}
-
-// GetWorkload gets the given workload
-func (client *Client) GetWorkload(workloadID string) (types.Workload, error) {
-	var wl types.Workload
-
-	url, err := client.getCiaoWorkloadsResource()
-	if err != nil {
-		return wl, errors.Wrap(err, "Error getting workloads resource")
-	}
-
-	url = fmt.Sprintf("%s/%s", url, workloadID)
-	err = client.getResource(url, api.WorkloadsV1, nil, &wl)
-
-	return wl, err
-}
-
-// ListComputeNodes returns the set of compute nodes
-func (client *Client) ListComputeNodes() (types.CiaoNodes, error) {
-	var nodes types.CiaoNodes
-
-	url := client.buildComputeURL("nodes/compute")
-	err := client.getResource(url, "", nil, &nodes)
-
-	return nodes, err
-}
-
-// ListNetworkNodes returns the set of network nodes
-func (client *Client) ListNetworkNodes() (types.CiaoNodes, error) {
-	var nodes types.CiaoNodes
-
-	url := client.buildComputeURL("nodes/network")
-	err := client.getResource(url, "", nil, &nodes)
-
-	return nodes, err
-}
-
-// ListNodes returns the set of network nodes
-func (client *Client) ListNodes() (types.CiaoNodes, error) {
-	var nodes types.CiaoNodes
-
-	url := client.buildComputeURL("nodes")
-	err := client.getResource(url, "", nil, &nodes)
-
-	return nodes, err
-}
-
-// ListCNCIs returns the set of CNCIs
-func (client *Client) ListCNCIs() (types.CiaoCNCIs, error) {
-	var nodes types.CiaoCNCIs
-
-	url := client.buildComputeURL("cncis")
-	err := client.getResource(url, "", nil, &nodes)
-
-	return nodes, err
-}
-
-// GetNodeSummary returns summary information about the cluster
-func (client *Client) GetNodeSummary() (types.CiaoClusterStatus, error) {
-	var status types.CiaoClusterStatus
-
-	url := client.buildComputeURL("nodes/summary")
-	err := client.getResource(url, "", nil, &status)
-
-	return status, err
-}
-
-// GetCNCI returns information about a CNCI
-func (client *Client) GetCNCI(cnciID string) (types.CiaoCNCI, error) {
-	var cnci types.CiaoCNCI
-
-	url := client.buildComputeURL("cncis/%s/detail", cnciID)
-	err := client.getResource(url, "", nil, &cnci)
-
-	return cnci, err
-}
-
-// ChangeNodeStatus modifies the status of a node
-func (client *Client) ChangeNodeStatus(nodeID string, status types.NodeStatusType) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	nodeStatus := types.CiaoNodeStatus{Status: status}
-
-	url, err := client.getCiaoResource("node", api.NodeV1)
-	if err != nil {
-		return errors.Wrap(err, "Error getting node resource")
-	}
-
-	url = fmt.Sprintf("%s/%s", url, nodeID)
-
-	err = client.putResource(url, api.NodeV1, &nodeStatus)
-
-	return err
-}
-
-func (client *Client) getCiaoQuotasResource() (string, error) {
-	return client.getCiaoResource("tenants", api.TenantsV1)
-}
-
-// UpdateQuotas updates the quotas for a given tenant
-func (client *Client) UpdateQuotas(tenantID string, quotas []types.QuotaDetails) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	url, err := client.getCiaoQuotasResource()
-	if err != nil {
-		return errors.Wrap(err, "Error getting quotas resource")
-	}
-
-	url = fmt.Sprintf("%s/%s/quotas", url, tenantID)
-	req := types.QuotaUpdateRequest{Quotas: quotas}
-	err = client.putResource(url, api.TenantsV1, &req)
-
-	return err
-}
-
-// ListQuotas lists the quotas for the specified tenant
-func (client *Client) ListQuotas(tenantID string) (types.QuotaListResponse, error) {
-	var result types.QuotaListResponse
-
-	url, err := client.getCiaoQuotasResource()
-	if err != nil {
-		return result, errors.Wrap(err, "Error getting quotas resource")
-	}
-
-	if tenantID != "" {
-		url = fmt.Sprintf("%s/%s/quotas", url, tenantID)
-	} else {
-		url = fmt.Sprintf("%s/quotas", url)
-	}
-
-	err = client.getResource(url, api.TenantsV1, nil, &result)
-
-	return result, err
-}
-
-func (client *Client) getCiaoTenantsResource() (string, error) {
-	url, err := client.getCiaoResource("tenants", api.TenantsV1)
-	return url, err
-}
-
-func (client *Client) getCiaoTenantRef(ID string) (string, error) {
-	var tenants types.TenantsListResponse
-
-	query := queryValue{
-		name:  "id",
-		value: ID,
-	}
-
-	url, err := client.getCiaoTenantsResource()
-	if err != nil {
-		return "", err
-	}
-
-	if !client.IsPrivileged() {
-		return url, errors.New("This command is only available to admins")
-	}
-
-	err = client.getResource(url, api.TenantsV1, []queryValue{query}, &tenants)
-	if err != nil {
-		return "", err
-	}
-
-	if len(tenants.Tenants) != 1 {
-		return "", errors.New("No tenant by that ID found")
-	}
-
-	links := tenants.Tenants[0].Links
-	url = client.getRef("self", links)
-	if url == "" {
-		return url, errors.New("invalid link returned from controller")
-	}
-
-	return url, nil
-}
-
-// GetTenantConfig gets the tenant configuration
-func (client *Client) GetTenantConfig(ID string) (types.TenantConfig, error) {
-	var config types.TenantConfig
-
-	url, err := client.getCiaoTenantRef(ID)
-	if err != nil {
-		return config, err
-	}
-
-	err = client.getResource(url, api.TenantsV1, nil, &config)
-
-	return config, err
-}
-
-// UpdateTenantConfig updates the tenant configuration
-func (client *Client) UpdateTenantConfig(ID string, name string, bits int) error {
-	var config types.TenantConfig
-
-	url, err := client.getCiaoTenantRef(ID)
-	if err != nil {
-		return err
-	}
-
-	err = client.getResource(url, api.TenantsV1, nil, &config)
-	if err != nil {
-		return err
-	}
-
-	oldconfig := config
-
-	if name != "" {
-		config.Name = name
-	}
-
-	if bits != 0 {
-		config.SubnetBits = bits
-	}
-
-	a, err := json.Marshal(oldconfig)
-	if err != nil {
-		return err
-	}
-
-	b, err := json.Marshal(config)
-	if err != nil {
-		return err
-	}
-
-	merge, err := jsonpatch.CreateMergePatch(a, b)
-	if err != nil {
-		return err
-	}
-
-	body := bytes.NewReader(merge)
-
-	_, err = client.sendHTTPRequest("PATCH", url, nil, body, "merge-patch+json")
-
-	return err
-}
-
-// CreateTenantConfig creates a new tenant configuration
-func (client *Client) CreateTenantConfig(tenantID string, name string, bits int) (types.TenantSummary, error) {
-	var req types.TenantRequest
-	var summary types.TenantSummary
-
-	if !client.IsPrivileged() {
-		return summary, errors.New("This command is only available to admins")
-	}
-
-	url, err := client.getCiaoTenantsResource()
-	if err != nil {
-		return summary, err
-	}
-
-	req.ID = tenantID
-	req.Config = types.TenantConfig{
-		Name:       name,
-		SubnetBits: bits,
-	}
-
-	err = client.postResource(url, api.TenantsV1, &req, &summary)
-
-	return summary, err
-}
-
-// DeleteTenant deletes the given tenant
-func (client *Client) DeleteTenant(tenantID string) error {
-	if !client.IsPrivileged() {
-		return errors.New("This command is only available to admins")
-	}
-
-	url, err := client.getCiaoTenantRef(tenantID)
-	if err != nil {
-		return err
-	}
-
-	return client.deleteResource(url, api.TenantsV1)
-}
-
-// ListTenantQuotas gets legacy tenant quota information
-func (client *Client) ListTenantQuotas() (types.CiaoTenantResources, error) {
-	var resources types.CiaoTenantResources
-
-	url := client.buildComputeURL("%s/quotas", client.TenantID)
-	err := client.getResource(url, "", nil, &resources)
-
-	return resources, err
-}
-
-// ListTenantResources gets tenant usage information
-func (client *Client) ListTenantResources() (types.CiaoUsageHistory, error) {
-	var usage types.CiaoUsageHistory
-	url := client.buildComputeURL("%s/resources", client.TenantID)
-
-	now := time.Now()
-	values := []queryValue{
-		{
-			name:  "start_date",
-			value: now.Add(-15 * time.Minute).Format(time.RFC3339),
-		},
-		{
-			name:  "end_date",
-			value: now.Format(time.RFC3339),
-		},
-	}
-
-	err := client.getResource(url, "", values, &usage)
-
-	return usage, err
-}
-
-// ListTenants returns a list of the tenants
-func (client *Client) ListTenants() (types.TenantsListResponse, error) {
-	var tenants types.TenantsListResponse
-
-	url, err := client.getCiaoTenantsResource()
-	if err != nil {
-		return tenants, err
-	}
-
-	err = client.getResource(url, api.TenantsV1, nil, &tenants)
-
-	return tenants, err
-}
-
-// ListTraceLabels returns a list of trace labels
-func (client *Client) ListTraceLabels() (types.CiaoTracesSummary, error) {
-	var traces types.CiaoTracesSummary
-
-	url := client.buildComputeURL("traces")
-	err := client.getResource(url, "", nil, &traces)
-
-	return traces, err
-}
-
-// GetTraceData returns trace details
-func (client *Client) GetTraceData(label string) (types.CiaoTraceData, error) {
-	var data types.CiaoTraceData
-
-	url := client.buildComputeURL("traces/%s", label)
-	err := client.getResource(url, "", nil, &data)
-
-	return data, err
-}
-
-// CreateVolume creates a volume from a request
-func (client *Client) CreateVolume(req api.RequestedVolume) (types.Volume, error) {
-	var vol types.Volume
-
-	url := client.buildCiaoURL("%s/volumes", client.TenantID)
-	err := client.postResource(url, api.VolumesV1, &req, &vol)
-
-	return vol, err
-}
-
-// ListVolumes lists the volumes
-func (client *Client) ListVolumes() ([]types.Volume, error) {
-	var volumes []types.Volume
-
-	url := client.buildCiaoURL("%s/volumes", client.TenantID)
-	err := client.getResource(url, api.VolumesV1, nil, &volumes)
-
-	return volumes, err
-}
-
-// GetVolume gets the details of a single volume
-func (client *Client) GetVolume(volumeID string) (types.Volume, error) {
-	var volume types.Volume
-
-	url := client.buildCiaoURL("%s/volumes/%s", client.TenantID, volumeID)
-	err := client.getResource(url, api.VolumesV1, nil, &volume)
-
-	return volume, err
-}
-
-// DeleteVolume deletes the volume
-func (client *Client) DeleteVolume(volumeID string) error {
-	url := client.buildCiaoURL("%s/volumes/%s", client.TenantID, volumeID)
-	return client.deleteResource(url, api.VolumesV1)
-}
-
-// AttachVolume attaches a volume to an instance
-func (client *Client) AttachVolume(volumeID string, instanceID, mountPoint string, mode string) error {
-	url := client.buildCiaoURL("%s/volumes/%s/action", client.TenantID, volumeID)
-
-	type AttachRequest struct {
-		MountPoint   string `json:"mountpoint"`
-		Mode         string `json:"mode"`
-		InstanceUUID string `json:"instance_uuid"`
-	}
-
-	// mountpoint or mode isn't required
-	var attachReq = struct {
-		Attach AttachRequest `json:"attach"`
-	}{
-		Attach: AttachRequest{
-			MountPoint:   mountPoint,
-			Mode:         mode,
-			InstanceUUID: instanceID,
-		},
-	}
-
-	err := client.postResource(url, api.VolumesV1, &attachReq, nil)
-
-	return err
-}
-
-// DetachVolume detaches a volume from an instances
-func (client *Client) DetachVolume(volumeID string) error {
-	url := client.buildCiaoURL("%s/volumes/%s/action", client.TenantID, volumeID)
-
-	type DetachRequest struct {
-		AttachmentID string `json:"attachment_id,omitempty"`
-	}
-	var detachReq = struct {
-		Detach DetachRequest `json:"detach"`
-	}{
-		Detach: DetachRequest{},
-	}
-
-	err := client.postResource(url, api.VolumesV1, &detachReq, nil)
-
-	return err
-
 }
